updated endpoint for checkout form, updated endpoint for getting order number

+++ server/public/api/orders.php
@@ -27,16 +27,28 @@ if ( $method === 'POST'){
     ];
   };
   $cartItem = json_encode( $cartItem );
+  // Create unique ID for order
+  $orderId = uniqid();
   // Add to order db
   $query = "INSERT INTO `orders` 
-  VALUES (null, '$_SESSION[user]','$cartItem', 'demo', 'demo', 
+  VALUES ('$orderId', '$_SESSION[user]','$cartItem', 'demo', 'demo', 
           'demo', 'demo', 'demo', 'CA', 'demo')";
   $output = mysqli_query( $conn , $query );
+
+  if ( !$output ){
+    throw new Exception( mysqli_error( $conn ));
+  }
 
   // Remove from cart db
   $deleteCart = "DELETE FROM `carts` WHERE user_id = '$_SESSION[user]'";
   $deleteOutput = mysqli_query( $conn, $deleteCart );
 
-  print_r( json_encode( ['success' => $output]) );
+  // Get Order ID
+  $orderIdQuery = "SELECT id FROM `orders`
+              WHERE id = '$orderId'";
+  $output = mysqli_query( $conn, $orderIdQuery );
+  $row = mysqli_fetch_assoc( $output );
+
+  print_r( json_encode( $row ));
   }
 ?>
